eet: be picky and really use the correct format without shitty cast.

+++ src/lib/eet/eet_lib.c
@@ -943,8 +943,8 @@ eet_internal_read2(Eet_File *ef)
         if (eet_test_close(!ef->ed, ef))
           return NULL;
 
-        INF("loading dictionnary for '%s' with %lu entries of size %lu",
-            ef->path, num_dictionary_entries, (unsigned long int) sizeof(Eet_String));
+        INF("loading dictionnary for '%s' with %lu entries of size %zu",
+            ef->path, num_dictionary_entries, sizeof(Eet_String));
 
         ef->ed->all = calloc(1, num_dictionary_entries * sizeof(Eet_String));
         if (eet_test_close(!ef->ed->all, ef))
